Adding daily batch of work.

+++ kotlin-multiplatform/src/jsMain/kotlin/ExtendedCalculatorJS.kt
@@ -1,6 +1,6 @@
 import kotlin.math.sqrt
 
 //@JsName("addAndSquareRoot")
-fun addAndSquareRoot(num1: Int, num2: Int): Double {
+fun addAndSquareRoot(num1: Double, num2: Double): Double {
     return sqrt(add(num1, num2))
 }
