+++ kotlin-multiplatform/src/jvmMain/kotlin/ExtendedCalculatorJVM.kt
@@ -1,5 +1,5 @@
 import kotlin.math.sqrt
 
-fun addAndSquareRoot(num1: Int, num2: Int): Double {
+fun addAndSquareRoot(num1: Double, num2: Double): Double {
     return sqrt(add(num1, num2))
 }
